PowerParams-class

Objects of PowerParams class store information required for sample size estimation for given data

Arguments

power_curve

1: Power simulation over a range of sample sizes or levels. 0: Power calculation over a single sample size or a level.

nsimn

The number of simulations to run. Default=1000

target_columns

Name of the experimental parameters to use for the power calculation.

levels

1: Amplify the number of corresponding target parameter. 0: Amplify the number of samples from the corresponding target parameter, ex) If target_columns = c("experiment","cell_line") and if you want to expand the number of experiment and sample more cells from each cell line, set levels = c(1,0).

max_size

Maximum levels or sample sizes to test. Default: the current level or the current sample size x 5. ex) If max_levels = c(10,5), it will test upto 10 experiments and 5 cell lines.

breaks

Levels /sample sizes of the variable to be specified along the power curve. Default: max(1, round( the number of current levels / 5 ))

effect_size

If you know the effect size of your condition variable, the effect size can be provided as a parameter. If the effect size is not provided, it will be estimated from your data

alpha

Threshold for Type I error

ICC

Intra-Class Coefficients (ICC) for each parameter

Value

an object of class ProbabilityModel

Examples

power_param=new("PowerParams")