calculatePower

This functions makes statistical power estimates given the data, the underlying design for it and the assumed probability model of the error distribution

Usage

calculatePower(data, design, model, power_param)

Arguments

data

Input data frame with columns having all the necessary information regarding the dependent and independent variables of interest

design

an object of class RMeDesign with the necessary design information about the data

model

an object of class ProbabilityModel giving the error distribution of the data

power_param

an object of class PowerParams giving the target parameter of interest and the other necessary parameter to perform the power estimation

Value

A power curve as a ggplot object or a power calculation result printed in a text file

Examples

result=calculatePower(data=data, design=design, model=model, power_param=power_param)
#> Error: object 'power_param' not found